Add first iteration of TBB-3.x Selenium WebDriver.
This was created to test the viability of running TBB-3.0b1 in a sandboxed
environment (provided by an 'experimental' compilation of pypy, which includes
the tracing just-in-time compiler and full-virtualisation sandboxing
features).  However, there appear to still be serious issues, and it is likely
that Selenium is not the tool we want to use to interract with a TorBrowser
instance, due to the fact that we cannot fully control XUL content windows via
Selenium.  See my comments and docstrings in the code for more details.